requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
boto3==1.28.4
botocore==1.31.4
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.2
daphne==4.0.0
distlib==0.3.6
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-environ==0.10.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
environ==1.0
filelock==3.12.2
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.1
multidict==6.0.4
mysqlclient==2.2.0
mysql-connector-python==8.0.33
openai==0.27.8
packaging==23.1
platformdirs==3.8.0
prompt-toolkit==3.0.39
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
yarl==1.9.2
zope.interface==6.0
django-storages==1.13.2
# django-prometheus

